:recycle: Undo rename av Modal.types for å unngå conflict
