+++ @navikt/core/react/src/modal/index.ts
@@ -2,4 +2,4 @@ export { default as Modal } from "./Modal";
 export { type ModalBodyProps } from "./ModalBody";
 export { type ModalFooterProps } from "./ModalFooter";
 export { type ModalHeaderProps } from "./ModalHeader";
-export { type ModalProps } from "./Modal.types";
+export { type ModalProps } from "./types";
